Move CommonPaths to Backend
