+++ src/Files.Uwp/Files.Uwp.csproj
@@ -189,7 +189,6 @@
     <Compile Include="Filesystem\BaseStorage\IBaseStorageFile.cs" />
     <Compile Include="Filesystem\BaseStorage\IBaseStorageFolder.cs" />
     <Compile Include="Helpers\ColorHelpers.cs" />
-    <Compile Include="Helpers\CommonPaths.cs" />
     <Compile Include="Helpers\FileExtensionHelpers.cs" />
     <Compile Include="Helpers\ItemListDisplayHelpers\BlockingListEnumerator.cs" />
     <Compile Include="Helpers\LocalizedEnumHelper.cs" />
